@@ -12,12 +12,21 @@
   "devDependencies": {
     "@types/node": "^14.14.21",
     "tslint": "^5.12.1",
-    "typescript": "^3.3.3"
+    "typescript": "^3.9.7"
   },
   "dependencies": {
     "amqplib": "^0.6.0",
     "bottlejs": "^2.0.0",
     "dotenv": "^8.2.0",
+    "email-deep-validator": "^3.3.0",
+    "handlebars": "^4.7.6",
+    "jsonwebtoken": "^8.5.1",
+    "mysql2": "^2.2.5",
+    "nodemailer": "^6.4.17",
+    "randomstring": "^1.1.5",
+    "redis": "^3.0.2",
+    "semver": "^7.3.4",
+    "sequelize": "^6.3.5",
     "uuid": "^8.3.2"
   }
 }
